requirements-api.txt
# VerbaAurea API 依赖包

# 基础依赖（继承自原项目）
beautifulsoup4==4.13.4
jieba==0.42.1
nltk==3.9.1
openpyxl==3.1.5
pandas==2.2.3
python_docx==1.1.2
xlrd==2.0.1
traits==7.0.2

# API框架依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 文件处理依赖
python-multipart==0.0.6
aiofiles==23.2.1

# 日志和监控
structlog==23.2.0
psutil==5.9.6

# 开发和测试依赖（可选）
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
